+++ lms/lms/doctype/sell_collateral_application/sell_collateral_application.py
@@ -4,7 +4,10 @@
 
 from __future__ import unicode_literals
 
+import json
+
 import frappe
+import requests
 import utils
 from frappe import _
 from frappe.model.document import Document
@@ -97,9 +100,6 @@ class SellCollateralApplication(Document):
 
         self.selling_collateral_value = 0
 
-        # price_map = {i.isin: i.price for i in self.items}
-        # sell_quantity_map = {i.isin: 0 for i in self.items}
-        # sell_requested_quantity_map = {i.isin: i.quantity for i in self.items}
         price_map = {
             "{}{}{}".format(
                 i.isin,
@@ -161,8 +161,6 @@ class SellCollateralApplication(Document):
             sell_quantity_map[isin_folio_combo] = (
                 sell_quantity_map[isin_folio_combo] + i.sell_quantity
             )
-            # i.price = price_map.get(i.isin)
-            # self.selling_collateral_value += i.sell_quantity * price_map.get(i.isin)
             if self.instrument_type == "Shares":
                 i.price = price_map.get(isin_folio_combo)
             else:
@@ -188,13 +186,6 @@ class SellCollateralApplication(Document):
                         i.quantity,
                     )
                 )
-        # for i in self.items:
-        #     if sell_quantity_map.get(i.isin) < i.quantity:
-        #         frappe.throw(
-        #             "You need to {} all {} of isin {}".format(
-        #                 applicaton_type, i.quantity, i.isin
-        #             )
-        #         )
 
     def before_submit(self):
         # check if all securities are sold
@@ -235,11 +226,7 @@ class SellCollateralApplication(Document):
                         "{}".format("-" + str(i.folio) if i.folio else ""),
                     )
                 )
-        """22-06-21 informed by vinayak"""
-        # if self.lender_selling_amount > self.selling_collateral_value:
-        #     frappe.throw(
-        #         "Can not sell amount more than {}".format(self.selling_collateral_value)
-        #     )
+
         if self.lender_selling_amount <= 0:
             frappe.throw("Please fix the Lender Selling Amount.")
 
@@ -263,14 +250,16 @@ class SellCollateralApplication(Document):
                     )
 
     def on_update(self):
-        msg_type = "sell collateral"
-        if self.instrument_type == "Mutual Fund":
-            msg_type = "invoke"
+        las_settings = frappe.get_single("LAS Settings")
 
         if self.status == "Rejected":
-            msg = "Dear Customer,\nSorry! Your {} request was turned down due to technical reasons. Please try again after sometime or reach out to us through 'Contact Us' on the app  -Spark Loans".format(
-                msg_type
-            )
+            if self.instrument_type == "Mutual Fund":
+                msg = "Dear Customer,\nSorry! Your invoke request was turned down due to technical reasons. You can reach out via the 'Contact Us' section of the app or please try again later using this link- {link} -Spark Loans".format(
+                    link=las_settings.my_securities
+                )
+
+            else:
+                msg = "Dear Customer,\nSorry! Your sell collateral request was turned down due to technical reasons. Please try again after sometime or reach out to us through 'Contact Us' on the app  -Spark Loans"
 
             receiver_list = [str(self.get_customer().phone)]
             if self.get_customer().get_kyc().mob_num:
@@ -333,7 +322,7 @@ class SellCollateralApplication(Document):
                     frappe.db.commit()
 
     def on_submit(self):
-
+        las_settings = frappe.get_single("LAS Settings")
         for i in self.sell_items:
             if i.sell_quantity > 0:
                 collateral_ledger_data = {
@@ -367,32 +356,11 @@ class SellCollateralApplication(Document):
 
         lender = self.get_lender()
 
-        # sell_collateral_charges = self.validate_loan_charges_amount(
-        #     lender,
-        #     sell_collateral_charges,
-        #     "sell_collateral_minimum_amount",
-        #     "sell_collateral_maximum_amount",
-        # )
-
-        # is_for_interest = False
-        # interest_entry = frappe.get_value(
-        #     "Loan Transaction",
-        #     {
-        #         "loan": self.loan,
-        #         "transaction_type": "Interest",
-        #         "unpaid_interest": [">", 0],
-        #     },
-        #     "name",
-        # )
-        # if interest_entry:
-        #     is_for_interest = True
         loan.create_loan_transaction(
             transaction_type="Sell Collateral",
             amount=self.lender_selling_amount,
-            # amount=self.selling_collateral_value,
             approve=True,
             loan_margin_shortfall_name=self.loan_margin_shortfall,
-            # is_for_interest=is_for_interest,
         )
 
         if self.instrument_type == "Shares":
@@ -433,39 +401,7 @@ class SellCollateralApplication(Document):
                         loan_margin_shortfall_name=self.loan_margin_shortfall,
                     )
                 )
-                # if lender.cgst_on_dp_reimbursementsell_charges > 0:
-                #     cgst = total_dp_reimburse_sell_charges * (
-                #         lender.cgst_on_dp_reimbursementsell_charges / 100
-                #     )
-                #     loan.create_loan_transaction(
-                #         transaction_type="CGST on DP Reimbursement(Sell) Charges",
-                #         amount=cgst,
-                #         gst_percent=lender.cgst_on_dp_reimbursementsell_charges,
-                #         charge_reference=total_dp_reimburse_sell_charges_reference.name,
-                #         approve=True,
-                #     )
-                # if lender.sgst_on_dp_reimbursementsell_charges > 0:
-                #     sgst = total_dp_reimburse_sell_charges * (
-                #         lender.sgst_on_dp_reimbursementsell_charges / 100
-                #     )
-                #     loan.create_loan_transaction(
-                #         transaction_type="SGST on DP Reimbursement(Sell) Charges",
-                #         amount=sgst,
-                #         gst_percent=lender.sgst_on_dp_reimbursementsell_charges,
-                #         charge_reference=total_dp_reimburse_sell_charges_reference.name,
-                #         approve=True,
-                #     )
-                # if lender.igst_on_dp_reimbursementsell_charges > 0:
-                #     igst = total_dp_reimburse_sell_charges * (
-                #         lender.igst_on_dp_reimbursementsell_charges / 100
-                #     )
-                #     loan.create_loan_transaction(
-                #         transaction_type="IGST on DP Reimbursement(Sell) Charges",
-                #         amount=igst,
-                #         gst_percent=lender.igst_on_dp_reimbursementsell_charges,
-                #         charge_reference=total_dp_reimburse_sell_charges_reference.name,
-                #         approve=True,
-                #     )
+
             if sell_collateral_charges > 0:
                 sell_collateral_charges_reference = loan.create_loan_transaction(
                     transaction_type="Sell Collateral Charges",
@@ -473,39 +409,7 @@ class SellCollateralApplication(Document):
                     approve=True,
                     loan_margin_shortfall_name=self.loan_margin_shortfall,
                 )
-                # if lender.cgst_on_sell_collateral_charges > 0:
-                #     cgst = sell_collateral_charges * (
-                #         lender.cgst_on_sell_collateral_charges / 100
-                #     )
-                #     loan.create_loan_transaction(
-                #         transaction_type="CGST on Sell Collateral Charges",
-                #         amount=cgst,
-                #         gst_percent=lender.cgst_on_sell_collateral_charges,
-                #         charge_reference=sell_collateral_charges_reference.name,
-                #         approve=True,
-                #     )
-                # if lender.sgst_on_sell_collateral_charges > 0:
-                #     sgst = sell_collateral_charges * (
-                #         lender.sgst_on_sell_collateral_charges / 100
-                #     )
-                #     loan.create_loan_transaction(
-                #         transaction_type="SGST on Sell Collateral Charges",
-                #         amount=sgst,
-                #         gst_percent=lender.sgst_on_sell_collateral_charges,
-                #         charge_reference=sell_collateral_charges_reference.name,
-                #         approve=True,
-                #     )
-                # if lender.igst_on_sell_collateral_charges > 0:
-                #     igst = sell_collateral_charges * (
-                #         lender.igst_on_sell_collateral_charges / 100
-                #     )
-                #     loan.create_loan_transaction(
-                #         transaction_type="IGST on Sell Collateral Charges",
-                #         amount=igst,
-                #         gst_percent=lender.igst_on_sell_collateral_charges,
-                #         charge_reference=sell_collateral_charges_reference.name,
-                #         approve=True,
-                #     )
+
         else:
             # invoke charges - Mutual Fund
             invoke_charges = lender.invoke_initiate_charges
@@ -529,33 +433,6 @@ class SellCollateralApplication(Document):
                     gst_percent=0,
                     loan_margin_shortfall_name=self.loan_margin_shortfall,
                 )
-                # if lender.cgst_on_invocation_charges > 0:
-                #     cgst = invoke_charges * (lender.cgst_on_invocation_charges / 100)
-                #     loan.create_loan_transaction(
-                #         transaction_type="CGST on Invocation Charges",
-                #         amount=cgst,
-                #         gst_percent=lender.cgst_on_invocation_charges,
-                #         charge_reference=invoke_charges_reference.name,
-                #         approve=True,
-                #     )
-                # if lender.sgst_on_invocation_charges > 0:
-                #     sgst = invoke_charges * (lender.sgst_on_invocation_charges / 100)
-                #     loan.create_loan_transaction(
-                #         transaction_type="SGST on Invocation Charges",
-                #         amount=sgst,
-                #         gst_percent=lender.sgst_on_invocation_charges,
-                #         charge_reference=invoke_charges_reference.name,
-                #         approve=True,
-                #     )
-                # if lender.igst_on_invocation_charges > 0:
-                #     igst = invoke_charges * (lender.igst_on_invocation_charges / 100)
-                #     loan.create_loan_transaction(
-                #         transaction_type="IGST on Invocation Charges",
-                #         amount=igst,
-                #         gst_percent=lender.igst_on_invocation_charges,
-                #         charge_reference=invoke_charges_reference.name,
-                #         approve=True,
-                #     )
 
         user_roles = frappe.db.get_values(
             "Has Role", {"parent": self.owner, "parenttype": "User"}, ["role"]
@@ -564,25 +441,26 @@ class SellCollateralApplication(Document):
             frappe.throw(_("Invalid User"))
         user_roles = [role[0] for role in user_roles]
 
-        msg_type = ["Sale of securities", "sale"]
         email_subject = "Sale Triggered Completion"
         application_type = "sell collateral"
         if loan.instrument_type == "Mutual Fund":
-            msg_type = ["Invoke", "invoke"]
             email_subject = "MF Sale Triggered Completion"
             application_type = "invoke"
 
         if "Loan customer" not in user_roles and self.loan_margin_shortfall:
             doc = frappe.get_doc("User KYC", self.get_customer().choice_kyc).as_dict()
             doc["sell_triggered_completion"] = {"loan": self.loan}
-            # if self.status in ["Pending", "Approved", "Rejected"]:
-            # email_subject = "Sale Triggered Completion"
-            # if self.instrument_type == "Mutual Fund":
-            # email_subject = "MF Sale Triggered Completion"
+
             frappe.enqueue_doc("Notification", email_subject, method="send", doc=doc)
-            msg = "Dear Customer,\n{} initiated by the lending partner for your loan account  {} is now completed .The {} proceeds have been credited to your loan account and collateral value updated. Please check the app for details. Spark Loans".format(
-                msg_type[0], self.loan, msg_type[1]
-            )
+            if loan.instrument_type == "Mutual Fund":
+                msg = "Dear Customer,\nInvoke initiated by the lending partner for your loan account  {} is now completed .The invoke proceeds have been credited to your loan account and collateral value updated. Please check the app for details - {link} -Spark Loans".format(
+                    self.loan, link=las_settings.my_loans
+                )
+            else:
+                msg = "Dear Customer,\nSale of securities initiated by the lending partner for your loan account  {} is now completed .The sale proceeds have been credited to your loan account and collateral value updated. Please check the app for details. Spark Loans".format(
+                    self.loan
+                )
+
             fcm_notification = frappe.get_doc(
                 "Spark Push Notification", "Sale triggerred completed", fields=["*"]
             )
@@ -593,15 +471,15 @@ class SellCollateralApplication(Document):
                 message = fcm_notification.message.format(Sale="Invoke", loan=self.loan)
                 fcm_notification = fcm_notification.as_dict()
                 fcm_notification["title"] = "Invoke triggerred completed "
-            # message = fcm_notification.message.format(loan=self.loan)
         else:
-            # msg_type = "sell collateral"
-            # if loan.instrument_type == "Mutual Fund":
-            #     msg_type = "invoke"
-
-            msg = "Dear Customer,\nCongratulations! Your {} request has been successfully executed and sale proceeds credited to your loan account. Kindly check the app for details -Spark Loans".format(
-                application_type
-            )
+            if loan.instrument_type == "Mutual Fund":
+                msg = "Dear Customer,\nCongratulations! Your {} request has been successfully executed and sale proceeds credited to your loan account. Kindly check the app for details - {link} -Spark Loans".format(
+                    application_type, link=las_settings.my_loans
+                )
+            else:
+                msg = "Dear Customer,\nCongratulations! Your {} request has been successfully executed and sale proceeds credited to your loan account. Kindly check the app for details -Spark Loans".format(
+                    application_type
+                )
 
             fcm_notification = frappe.get_doc(
                 "Spark Push Notification", "Sell request executed", fields=["*"]
@@ -621,7 +499,7 @@ class SellCollateralApplication(Document):
             receiver_list = list(set(receiver_list))
 
             frappe.enqueue(method=send_sms, receiver_list=receiver_list, msg=msg)
-        # loan.update_loan_balance()
+
         self.notify_customer(fcm_notification=fcm_notification, message=message)
 
     def validate(self):
@@ -677,11 +555,6 @@ def get_collateral_details(sell_collateral_application_name):
         ),
         having_clause=" HAVING quantity > 0",
     )
-
-
-import json
-
-import requests
 
 
 @frappe.whitelist()
@@ -745,6 +618,12 @@ def validate_invoc(sell_collateral_application_name):
                     )
                     data["invocvalidate"]["schemedetails"].append(schemedetails[0])
 
+                lms.create_log(
+                    {
+                        "json_payload": data,
+                    },
+                    "invoke_validate_request",
+                )
                 encrypted_data = lms.AESCBC(
                     las_settings.encryption_key, las_settings.iv
                 ).encrypt(json.dumps(data))
@@ -765,12 +644,11 @@ def validate_invoc(sell_collateral_application_name):
 
                 lms.create_log(
                     {
-                        "json_payload": data,
                         "encrypted_request": encrypted_data,
                         "encrypred_response": json.loads(resp).get("res"),
                         "decrypted_response": dict_decrypted_response,
                     },
-                    "invoke_validate",
+                    "invoke_validate_response",
                 )
 
                 if dict_decrypted_response.get("invocvalidate"):
@@ -886,11 +764,21 @@ def initiate_invoc(sell_collateral_application_name):
                     )
                     data["invocinitiate"]["schemedetails"].append(schemedetails[0])
 
+                lms.create_log(
+                    {
+                        "json_payload": data,
+                    },
+                    "invoke_initiate_request",
+                )
                 encrypted_data = lms.AESCBC(
                     las_settings.encryption_key, las_settings.iv
                 ).encrypt(json.dumps(data))
 
                 req_data = {"req": str(encrypted_data)}
+
+                sell_collateral_application_doc.db_set(
+                    "invoke_initiate_request_timestamp", frappe.utils.now_datetime()
+                )
 
                 resp = requests.post(
                     url=url, headers=headers, data=json.dumps(req_data)
@@ -906,12 +794,11 @@ def initiate_invoc(sell_collateral_application_name):
 
                 lms.create_log(
                     {
-                        "json_payload": data,
                         "encrypted_request": encrypted_data,
                         "encrypred_response": json.loads(resp).get("res"),
                         "decrypted_response": dict_decrypted_response,
                     },
-                    "invoke_initiate",
+                    "invoke_initiate_response",
                 )
 
                 if dict_decrypted_response.get("invocinitiate"):
@@ -977,6 +864,12 @@ def initiate_invoc(sell_collateral_application_name):
         else:
             frappe.throw(frappe._("Mycams Email ID is missing"))
     except utils.exceptions.APIException as e:
+        sell_collateral_application_doc = frappe.get_doc(
+            "Sell Collateral Application", sell_collateral_application_name
+        )
+        sell_collateral_application_doc.db_set(
+            "invoke_initiate_request_timestamp", frappe.utils.now_datetime()
+        )
         frappe.log_error(
             title="Invocation - Initiate - Error",
             message=frappe.get_traceback()