+++ lms/lms/doctype/top_up_application/top_up_application.py
@@ -366,12 +366,21 @@ class TopupApplication(Document):
                 "Customer Address Details", user_kyc.address_details
             )
             address = (
-                str(address_details.perm_line1)
-                + ", "
-                + str(address_details.perm_line2)
-                + ", "
-                + str(address_details.perm_line3)
-                + ", "
+                (
+                    (str(address_details.perm_line1) + ", ")
+                    if address_details.perm_line1
+                    else ""
+                )
+                + (
+                    (str(address_details.perm_line2) + ", ")
+                    if address_details.perm_line2
+                    else ""
+                )
+                + (
+                    (str(address_details.perm_line3) + ", ")
+                    if address_details.perm_line3
+                    else ""
+                )
                 + str(address_details.perm_city)
                 + ", "
                 + str(address_details.perm_dist)