+++ lms/lms/doctype/unpledge_application/unpledge_application.py
@@ -4,7 +4,10 @@
 
 from __future__ import unicode_literals
 
+import json
+
 import frappe
+import requests
 import utils
 from frappe.model.document import Document
 
@@ -110,9 +113,6 @@ class UnpledgeApplication(Document):
 
         application_type = "unpledge" if self.instrument_type == "Shares" else "revoke"
 
-        # price_map = {i.isin: i.price for i in self.items}
-        # unpledge_quantity_map = {i.isin: 0 for i in self.items}
-        # unpledge_requested_quantity_map = {i.isin: i.quantity for i in self.items}
         price_map = {
             "{}{}{}".format(
                 i.isin,
@@ -176,10 +176,7 @@ class UnpledgeApplication(Document):
             unpledge_quantity_map[isin_folio_combo] = (
                 unpledge_quantity_map[isin_folio_combo] + i.unpledge_quantity
             )
-            # i.price = price_map.get(i.isin)
-            # self.unpledge_collateral_value += i.unpledge_quantity * price_map.get(
-            #     i.isin
-            # )
+
             if self.instrument_type == "Shares":
                 i.price = price_map.get(isin_folio_combo)
             else:
@@ -202,13 +199,6 @@ class UnpledgeApplication(Document):
                         application_type, isin_folio_combo, i.quantity
                     )
                 )
-        # for i in self.items:
-        #     if unpledge_quantity_map.get(i.isin) < i.quantity:
-        #         frappe.throw(
-        #             "You need to {} all {} of isin {}".format(
-        #                 application_type, i.quantity, i.isin
-        #             )
-        #         )
 
     def before_submit(self):
         # check if all securities are sold
@@ -242,7 +232,6 @@ class UnpledgeApplication(Document):
                 "{}".format("-" + str(i.folio) if i.folio else ""),
                 i.psn if i.psn else "",
             )
-            # print(unpledge_quantity_map.get(i.isin), i.quantity)
             if unpledge_quantity_map.get(isin_folio_combo) < i.quantity:
                 frappe.throw(
                     "You need to {} all {} of isin {}".format(
@@ -325,39 +314,7 @@ class UnpledgeApplication(Document):
                         approve=True,
                     )
                 )
-                # if lender.cgst_on_dp_reimbursementunpledge_charges > 0:
-                #     cgst = total_dp_reimburse_unpledge_charges * (
-                #         lender.cgst_on_dp_reimbursementunpledge_charges / 100
-                #     )
-                #     loan.create_loan_transaction(
-                #         transaction_type="CGST on DP Reimbursement(Unpledge) Charges",
-                #         amount=cgst,
-                #         gst_percent=lender.cgst_on_dp_reimbursementunpledge_charges,
-                #         charge_reference=total_dp_reimburse_unpledge_charges_reference.name,
-                #         approve=True,
-                #     )
-                # if lender.sgst_on_dp_reimbursementunpledge_charges > 0:
-                #     sgst = total_dp_reimburse_unpledge_charges * (
-                #         lender.sgst_on_dp_reimbursementunpledge_charges / 100
-                #     )
-                #     loan.create_loan_transaction(
-                #         transaction_type="SGST on DP Reimbursement(Unpledge) Charges",
-                #         amount=sgst,
-                #         gst_percent=lender.sgst_on_dp_reimbursementunpledge_charges,
-                #         charge_reference=total_dp_reimburse_unpledge_charges_reference.name,
-                #         approve=True,
-                #     )
-                # if lender.igst_on_dp_reimbursementunpledge_charges > 0:
-                #     igst = total_dp_reimburse_unpledge_charges * (
-                #         lender.igst_on_dp_reimbursementunpledge_charges / 100
-                #     )
-                #     loan.create_loan_transaction(
-                #         transaction_type="IGST on DP Reimbursement(Unpledge) Charges",
-                #         amount=igst,
-                #         gst_percent=lender.igst_on_dp_reimbursementunpledge_charges,
-                #         charge_reference=total_dp_reimburse_unpledge_charges_reference.name,
-                #         approve=True,
-                #     )
+
         else:
             # revoke charges - Mutual Fund
             revoke_charges = lender.revoke_initiate_charges
@@ -379,37 +336,11 @@ class UnpledgeApplication(Document):
                     amount=revoke_charges,
                     approve=True,
                 )
-                # if lender.cgst_on_revocation_charges > 0:
-                #     cgst = revoke_charges * (lender.cgst_on_revocation_charges / 100)
-                #     loan.create_loan_transaction(
-                #         transaction_type="CGST on Revocation Charges",
-                #         amount=cgst,
-                #         gst_percent=lender.cgst_on_revocation_charges,
-                #         charge_reference=revoke_charges_reference.name,
-                #         approve=True,
-                #     )
-                # if lender.sgst_on_revocation_charges > 0:
-                #     sgst = revoke_charges * (lender.sgst_on_revocation_charges / 100)
-                #     loan.create_loan_transaction(
-                #         transaction_type="SGST on Revocation Charges",
-                #         amount=sgst,
-                #         gst_percent=lender.sgst_on_revocation_charges,
-                #         charge_reference=revoke_charges_reference.name,
-                #         approve=True,
-                #     )
-                # if lender.igst_on_revocation_charges > 0:
-                #     igst = revoke_charges * (lender.igst_on_revocation_charges / 100)
-                #     loan.create_loan_transaction(
-                #         transaction_type="IGST on Revocation Charges",
-                #         amount=igst,
-                #         gst_percent=lender.igst_on_revocation_charges,
-                #         charge_reference=revoke_charges_reference.name,
-                #         approve=True,
-                #     )
 
         self.notify_customer()
 
     def notify_customer(self, check=None):
+        las_settings = frappe.get_single("LAS Settings")
         msg = ""
         fcm_notification = {}
 
@@ -441,19 +372,19 @@ class UnpledgeApplication(Document):
                     fcm_notification["title"] = "Revoke application accepted"
             elif self.status == "Rejected":
                 if check == True:
-                    # msg = """Dear {},
-                    # Your unpledge request was rejected.
-                    # There is a margin shortfall.
-                    # You can send another unpledge request when there is no margin shortfall.""".format(
-                    #     self.get_loan().get_customer().first_name
-                    # )
                     msg = """Your {msg_type} request was rejected. There is a margin shortfall. You can send another {msg_type} request when there is no margin shortfall.""".format(
                         msg_type=msg_type
                     )
                 else:
-                    msg = "Dear Customer,\nSorry! Your {} application was turned down due to technical reasons. Please try again after sometime or reach us through 'Contact Us' on the app  -Spark Loans".format(
-                        msg_type
-                    )
+                    if self.instrument_type == "Mutual Fund":
+                        msg = "Dear Customer,\nSorry! Your {} application was turned down due to technical reasons. You can reach out via the 'Contact Us' section of the app or please try again later using this link- {link} -Spark Loans".format(
+                            msg_type, link=las_settings.my_securities
+                        )
+                    else:
+                        msg = "Dear Customer,\nSorry! Your {} application was turned down due to technical reasons. Please try again after sometime or reach us through 'Contact Us' on the app  -Spark Loans".format(
+                            msg_type
+                        )
+
                     fcm_notification = frappe.get_doc(
                         "Spark Push Notification",
                         "Unpledge application rejected",
@@ -507,19 +438,6 @@ class UnpledgeApplication(Document):
     def get_lender(self):
         return frappe.get_doc("Lender", self.lender)
 
-    # def check(self):
-    #     loan = self.get_loan()
-    #     final_drawing_power = (loan.total_collateral_value - self.unpledge_collateral_value)*loan.allowable_ltv/100
-    #     print(loan.total_collateral_value,"loan.total_collateral_value")
-    #     print(self.unpledge_collateral_value,"self.unpledge_collateral_value")
-    #     print(final_drawing_power,"final_drawing_power")
-    #     if self.status == "Pending" and final_drawing_power < loan.balance:
-    #         self.status = "Rejected"
-    #         self.workflow_state = "Rejected"
-    #         self.save(ignore_permissions=True)
-    #         frappe.db.commit()
-    #         self.notify_customer(check=True)
-
 
 @frappe.whitelist()
 def get_collateral_details(unpledge_application_name):
@@ -548,11 +466,6 @@ def get_collateral_details(unpledge_application_name):
         ),
         having_clause=" HAVING quantity > 0",
     )
-
-
-import json
-
-import requests
 
 
 @frappe.whitelist()
@@ -614,6 +527,12 @@ def validate_revoc(unpledge_application_name):
                     )
                     data["revocvalidate"]["schemedetails"].append(schemedetails[0])
 
+                lms.create_log(
+                    {
+                        "json_payload": data,
+                    },
+                    "revoke_validate_request",
+                )
                 encrypted_data = lms.AESCBC(
                     las_settings.encryption_key, las_settings.iv
                 ).encrypt(json.dumps(data))
@@ -634,12 +553,11 @@ def validate_revoc(unpledge_application_name):
 
                 lms.create_log(
                     {
-                        "json_payload": data,
                         "encrypted_request": encrypted_data,
                         "encrypred_response": json.loads(resp).get("res"),
                         "decrypted_response": dict_decrypted_response,
                     },
-                    "revoke_validate",
+                    "revoke_validate_response",
                 )
 
                 if dict_decrypted_response.get("revocvalidate"):
@@ -755,6 +673,12 @@ def initiate_revoc(unpledge_application_name):
                     )
                     data["revocinitiate"]["schemedetails"].append(schemedetails[0])
 
+                lms.create_log(
+                    {
+                        "json_payload": data,
+                    },
+                    "revoke_initiate_request",
+                )
                 encrypted_data = lms.AESCBC(
                     las_settings.encryption_key, las_settings.iv
                 ).encrypt(json.dumps(data))
@@ -775,12 +699,11 @@ def initiate_revoc(unpledge_application_name):
 
                 lms.create_log(
                     {
-                        "json_payload": data,
                         "encrypted_request": encrypted_data,
                         "encrypred_response": json.loads(resp).get("res"),
                         "decrypted_response": dict_decrypted_response,
                     },
-                    "revoke_initiate",
+                    "revoke_initiate_response",
                 )
 
                 if dict_decrypted_response.get("revocinitiate"):