+++ lms/lms/doctype/user_token/user_token.py
@@ -105,51 +105,6 @@ class UserToken(Document):
             #     app_hash_string=app_hash_string,
             #     expiry_in_minutes=expiry_in_minutes,
             # )
-            token_type = self.token_type.replace(" ", "")
-            if self.token_type in [
-                "Pledge OTP",
-                "Unpledge OTP",
-            ]:
-                doc = frappe.get_all(
-                    "User KYC", filters={"user": frappe.session.user}, fields=["*"]
-                )[0]
-                email_otp = frappe.db.sql(
-                    "select message from `tabNotification` where name='OTP for Spark Loans';"
-                )[0][0]
-                email_otp = email_otp.replace("investor_name", doc.fullname)
-                email_otp = email_otp.replace("token_type", token_type)
-                email_otp = email_otp.replace("token", self.token)
-                frappe.enqueue(
-                    method=frappe.sendmail,
-                    recipients=[doc.email],
-                    sender=None,
-                    subject="OTP for Spark Loans",
-                    message=email_otp,
-                    queue="short",
-                    delayed=False,
-                    job_name="Spark OTP on Email",
-                )
-            else:
-                doc = frappe.get_all(
-                    "User", filters={"phone": self.entity}, fields=["*"]
-                )
-                if doc:
-                    email_otp = frappe.db.sql(
-                        "select message from `tabNotification` where name='Other OTP for Spark Loans';"
-                    )[0][0]
-                    email_otp = email_otp.replace("investor_name", doc[0].full_name)
-                    email_otp = email_otp.replace("token_type", token_type)
-                    email_otp = email_otp.replace("token", self.token)
-                    frappe.enqueue(
-                        method=frappe.sendmail,
-                        recipients=[doc[0].email],
-                        sender=None,
-                        subject="OTP for Spark Loans",
-                        message=email_otp,
-                        queue="short",
-                        delayed=False,
-                        job_name="Spark OTP on Email",
-                    )
 
             mess = frappe._(
                 "Dear Customer, Your {token_type} for Spark Loans is {token}. Do not share your {token_type} with anyone. Your OTP is valid for 10 minutes -Spark Loans"
@@ -285,56 +240,66 @@ def validate_receiver_nos(receiver_list):
 
 
 def send_sms(receiver_list, msg, sender_name="", success_msg=True):
+    arg = {}
+    try:
 
-    import json
+        import json
 
-    from six import string_types
+        from six import string_types
 
-    if isinstance(receiver_list, string_types):
-        receiver_list = json.loads(receiver_list)
-        if not isinstance(receiver_list, list):
-            receiver_list = [receiver_list]
+        if isinstance(receiver_list, string_types):
+            receiver_list = json.loads(receiver_list)
+            if not isinstance(receiver_list, list):
+                receiver_list = [receiver_list]
 
-    receiver_list = validate_receiver_nos(receiver_list)
+        receiver_list = validate_receiver_nos(receiver_list)
 
-    arg = {
-        "receiver_list": receiver_list,
-        "message": frappe.safe_decode(msg).encode("utf-8"),
-        "success_msg": success_msg,
-    }
+        arg = {
+            "receiver_list": receiver_list,
+            "message": frappe.safe_decode(msg).encode("utf-8"),
+            "success_msg": success_msg,
+        }
 
-    if frappe.db.get_value("SMS Settings", None, "sms_gateway_url"):
-        send_via_gateway(arg)
-    else:
-        frappe.msgprint(_("Please Update SMS Settings"))
+        if frappe.db.get_value("SMS Settings", None, "sms_gateway_url"):
+            send_via_gateway(arg)
+        else:
+            frappe.msgprint(_("Please Update SMS Settings"))
+    except Exception:
+        frappe.log_error(
+            message=frappe.get_traceback() + "\n\nArguments:\n" + str(arg),
+            title="Send SMS Error",
+        )
 
 
 def send_via_gateway(arg):
-    ss = frappe.get_doc("SMS Settings", "SMS Settings")
-    headers = get_headers(ss)
+    try:
+        ss = frappe.get_doc("SMS Settings", "SMS Settings")
+        headers = get_headers(ss)
 
-    args = {ss.message_parameter: arg.get("message")}
-    for d in ss.get("parameters"):
-        if not d.header:
-            args[d.parameter] = d.value
+        args = {ss.message_parameter: arg.get("message")}
+        for d in ss.get("parameters"):
+            if not d.header:
+                args[d.parameter] = d.value
 
-    success_list = []
-    for d in arg.get("receiver_list"):
-        args[ss.receiver_parameter] = d
-        status = send_request(ss.sms_gateway_url, args, headers, ss.use_post)
+        success_list = []
+        for d in arg.get("receiver_list"):
+            args[ss.receiver_parameter] = d
+            status = send_request(ss.sms_gateway_url, args, headers, ss.use_post)
 
-        if 200 <= status < 300:
-            success_list.append(d)
+            if 200 <= status < 300:
+                success_list.append(d)
 
-    if len(success_list) > 0:
-        args.update(arg)
-        create_sms_log(args, success_list)
-        if arg.get("success_msg"):
-            frappe.msgprint(
-                frappe._("SMS sent to following numbers: {0}").format(
-                    "\n" + "\n".join(success_list)
+        if len(success_list) > 0:
+            args.update(arg)
+            create_sms_log(args, success_list)
+            if arg.get("success_msg"):
+                frappe.msgprint(
+                    frappe._("SMS sent to following numbers: {0}").format(
+                        "\n" + "\n".join(success_list)
+                    )
                 )
-            )
+    except Exception:
+        frappe.log_error()
 
 
 def get_headers(sms_settings=None):
@@ -350,30 +315,39 @@ def get_headers(sms_settings=None):
 
 
 def send_request(gateway_url, params, headers=None, use_post=False):
-    import requests
+    log = {}
+    try:
+        import requests
 
-    if not headers:
-        headers = get_headers()
+        if not headers:
+            headers = get_headers()
 
-    if use_post:
-        response = requests.post(gateway_url, headers=headers, data=params)
-    else:
-        response = requests.get(gateway_url, headers=headers, params=params)
-    # SMS LOG
-    import json
+        if use_post:
+            response = requests.post(gateway_url, headers=headers, data=params)
+        else:
+            response = requests.get(gateway_url, headers=headers, params=params)
+        # SMS LOG
+        import json
 
-    frappe.logger().info(params)
-    if type(params["sms"]) == bytes:
-        params["sms"] = params["sms"].decode("ascii")
-    log = {
-        "url": gateway_url,
-        "params": params,
-        "response": response.json(),
-    }
-    lms.create_log(log, "sms_log")
-    # SMS LOG end
-    response.raise_for_status()
-    return response.status_code
+        frappe.logger().info(params)
+        if type(params["sms"]) == bytes:
+            params["sms"] = params["sms"].decode("ascii")
+        log = {
+            "url": gateway_url,
+            "params": params,
+            "response": response.json(),
+        }
+        lms.create_log(log, "sms_log")
+        # SMS LOG end
+        response.raise_for_status()
+        return response.status_code
+    except Exception:
+        frappe.log_error(
+            message=frappe.get_traceback() + "\n\nSMS details:\n" + str(log)
+            if log
+            else params,
+            title="Send sms request error",
+        )
 
 
 # Create SMS Log